Various changes
Threaded backups. This means that the server will not freeze anymore
while the files are beung copied. Theorically, no map corruptions will
occur neither
Multiple lines messages in the config is now supported for all sections.
Use the "<n>" tag to break the line.
Minor fixes in the config.
Code improvement in the PlayerJoinEvent handling.